- some leftover shorts plus utilities that ended up unused but may be useful later.

+++ source/games/duke/src/funct.h
@@ -136,7 +136,7 @@ int madenoise(int playerNum);
 int haskey(int sect, int snum);
 void shootbloodsplat(DDukeActor* i, int p, int sx, int sy, int sz, int sa, int atwith, int BIGFORCE, int OOZFILTER, int NEWBEAST);
 
-void breakwall(short newpn, DDukeActor* spr, short dawallnum);
+void breakwall(int newpn, DDukeActor* spr, int dawallnum);
 int callsound(int sectnum,DDukeActor* snum);
 int hitasprite(DDukeActor* snum,DDukeActor **hitSprite);
 int findplayer(const DDukeActor* s, int* dist);
